Consolidate NewsLayout tests and remove things that aren’t being used
# Conflicts:
#	src/Components/Publishing/Layouts/__tests__/__snapshots__/NewsLayoutAdsEnabled.test.tsx.snap
